Move IcmpTypes definitions into top level Icmp
