+++ clash-eth.cabal
@@ -102,7 +102,6 @@ library
     Clash.Cores.Ethernet.Examples.EchoStack
     Clash.Cores.Ethernet.Examples.RxStack
     Clash.Cores.Ethernet.Examples.TxStack
-    Clash.Cores.Ethernet.Icmp.IcmpTypes
     Clash.Cores.Ethernet.IP.InternetChecksum
     Clash.Cores.Ethernet.IP.IPPacketizers
     Clash.Cores.Ethernet.IP.IPv4Types
